add netlify github action (#5)
This workflow deploys the dev branch to Netlify when we add code. This way, we have a branch for code not ready for production.

It auto runs when a PR is merged into dev. It can also be triggered manually.

If auto triggered, a comment is added to the PR with the URL. If run manually, no comment is added.